Add Kaiju token
The Project Kaiju is created for 2 things assist of  [NFT  Game : Kaiju Card Battle]  and [NFT Kaiju Collected & Marketplace Kaiju Card Battle & Collected] is a new NFT card game. Player can buy and sell card in Kaiju market. All card selling by system are in the random box with the difference rarity including with Normal, Rare, UltraRare,SecretRare,Legendary. The card in the first design is in the img below. One box include 5 cards inside. The player have to buy minimum 20 cards for battle with NPC of Kaiju card battle. But if the player want to PVP , player have to buy minimum 30 cards. All player have a rank base on stat of win/lose : No Rank, Baby Kaiju, Great Kaiju, Uni Kaiju, Legendary Kaiju. How to win (in draf) ? all cards are including with Elementary and Attack/Defend point. Attack not is everything because affect of element in important too. The player must pay fee for playing (~ 1 kanji, base on the price) for playing with NFC and got Kaiju for the reward if player win. If the player need to PVP. The player must pay fee for playing (~ 1 kanji, base on the price) and enter Kaiju coin for wager 5-500 Kaiju base on rank of player. The winner will got all wager kanji into winer wallet.